[pyroSAR.examine] new module for software examination handling
* instantiates ConfigHandler, instead of __init__
* accommodates classes ExamineGamma and ExamineSnap
