+++ pyroSAR/__init__.py
@@ -1,7 +1,4 @@
 from .drivers import *
-from ._dev_config import ConfigHandler
-
-ConfigHandler = ConfigHandler()
 
 from . import ancillary, drivers
