+++ pyroSAR/gamma/__init__.py
@@ -1,3 +1,3 @@
-from .auxil import process, ISPPar, UTM, Spacing, Namespace, slc_corners, ExamineGamma, par2hdr
+from .auxil import process, ISPPar, UTM, Spacing, Namespace, slc_corners, par2hdr
 from .util import geocode, multilook, ovs, convert2gamma, calibrate, correctOSV, S1_deburst
 from . import dem
