+++ pyroSAR/gamma/auxil.py
@@ -5,17 +5,15 @@
 import math
 import os
 import re
-import json
 import string
-import shutil
 import codecs
 import subprocess as sp
 from datetime import datetime
 
+from pyroSAR.examine import ExamineGamma
 from spatialist.ancillary import parse_literal, run, union, dissolve
 from spatialist.envi import hdr
 
-from pyroSAR import ConfigHandler
 from .error import gammaErrorHandler
 
 
@@ -424,45 +422,3 @@ def slc_corners(parfile):
             pts['xmin'], pts['xmax'] = [float(x) for x in
                                         re.findall(r'[0-9]+\.[0-9]+', line)]
     return pts
-
-
-class ExamineGamma(object):
-    def __init__(self):
-        home_sys = os.environ.get('GAMMA_HOME')
-        if 'GAMMA' in ConfigHandler.sections:
-            attr = ConfigHandler['GAMMA']
-            for key, value in attr.items():
-                setattr(self, key, value)
-        if hasattr(self, 'home'):
-            if home_sys is not None and self.home != home_sys:
-                print('the value of GAMMA_HOME is different to that in the pyroSAR configuration;\n'
-                      '  was: {}\n'
-                      '  is : {}\n'
-                      'resetting the configuration and deleting parsed modules'
-                      .format(self.home, home_sys))
-                parsed = os.path.join(os.path.dirname(ConfigHandler.file), 'gammaparse')
-                shutil.rmtree(parsed)
-                self.home = home_sys
-        if not hasattr(self, 'home'):
-            if home_sys is not None:
-                setattr(self, 'home', home_sys)
-            else:
-                raise RuntimeError('could not read Gamma installation directory')
-        self.version = re.search('GAMMA_SOFTWARE-(?P<version>[0-9]{8})',
-                                 getattr(self, 'home')).group('version')
-        self.__update_config()
-    
-    def __update_config(self):
-        if 'GAMMA' not in ConfigHandler.sections:
-            ConfigHandler.add_section('GAMMA')
-        
-        for attr in ['home', 'version']:
-            self.__update_config_attr(attr, getattr(self, attr), 'GAMMA')
-    
-    @staticmethod
-    def __update_config_attr(attr, value, section):
-        if isinstance(value, list):
-            value = json.dumps(value)
-        
-        if attr not in ConfigHandler[section].keys() or ConfigHandler[section][attr] != value:
-            ConfigHandler.set(section, key=attr, value=value, overwrite=True)
