+++ pyroSAR/snap/__init__.py
@@ -1,3 +1,2 @@
 from .util import geocode
 from .auxil import gpt
-from pyroSAR._dev_config import ExamineSnap
