+++ pyroSAR/tests/test_snap.py
@@ -5,7 +5,7 @@ from pyroSAR.snap import geocode
 from spatialist import bbox
 from spatialist.ancillary import finder
 from pyroSAR.snap.auxil import is_consistent, split, groupbyWorkers, Workflow, parse_recipe
-from pyroSAR._dev_config import ExamineSnap
+from pyroSAR.examine import ExamineSnap
 
 
 def test_installation():
